Move stuff from factory to pydantic models
Specifically, move all the "helper"/utility functionalities from
`datasource_factory.py` to a new `soft7_instance.py` under
pydantic_models.
