+++ .pre-commit-config.yaml
@@ -15,6 +15,11 @@ repos:
     - id: trailing-whitespace
       args: [--markdown-linebreak-ext=md]
 
+  - repo: https://github.com/ambv/black
+    rev: 23.9.1
+    hooks:
+    - id: black
+
   - repo: https://github.com/astral-sh/ruff-pre-commit
     rev: v0.0.292
     hooks:
@@ -22,11 +27,6 @@ repos:
         args:
         - "--fix"
         - "--exit-non-zero-on-fix"
-
-  - repo: https://github.com/ambv/black
-    rev: 23.9.1
-    hooks:
-    - id: black
 
   - repo: https://github.com/PyCQA/bandit
     rev: 1.7.5
