+++ s7/pydantic_models/soft7_entity.py
